feat(THA-1146): implement basic vetoken (#153)
In the PR, we build`vetoken` by learning from `0x1::coin` interface.

## Algorithm

The most important feature of a VeToken is that its strength decays over
time, and we care about the total_strength (= sum of strength across all
minted vetoken).

We use `locked_amounts` vector to store the locked amount that is about
to unlock in 1, 2, 3, ..., `max_duration_weeks` weeks. By simply
iterating through the vector we're able to calculate total_strength of
the specific vetoken (for example VeTHL) in O(max_duration_weeks) time
complexity instead of having to go through each and every VeToken
issuance, which gonna be O(n) where n = number of VeToken owners.

In order to keep `locked_amounts` up-to-date, we `checkpoint` on every
vetoken interactions.

## TODOs

Here's only the start. There are many things todos to finish up the
vetoken module:

- [x] tests
- [ ] implement `lock_more`
- [ ] implement `extend_lock`
- [ ] integrate with parliament regarding voting power
- [ ] integrate with vesting regarding fee distribution
- [ ] evaluate aptos token v2 and see if we can migrate