[vetoken] never purge total locked amounts (#165)
In order to accurately compute past weights, we need to also need to
query for locked amounts for any week's epoch

1. Switch locked_amounts from `vector` to `Table` where the key is the
week epoch. We use a table as there may be a week where no amounts were
locked.
    - No need to have `fun checkpoint` which purges expired entries
    - Simplifies logic when extending the lock duration
2. Represent locked amounts as u128 as aggregate supply may exceed u64
3. When a token is locked, express duration as expected end week instead
of `duration_weeks`. Since our granularity is on a week-by-week basis,
the user should specify the epoch in which they expect to unlock and not
a total duration which cant be guaranteed
4. Fix up error names and add documentation. When locking/extending, we
should assert that there is a VeToken that is currently locked


NOTE:
- Since we have a week-by-week snapshot granularity, there's no way to
distinguish timestamps within a given epoch. Since a user acquires a
balance immediately when locking, this implies that the user also
retroactively obtains a balance in the past for that given epoch.
- If we wish to create a more "fair" system and prevent this, the user
should acquire a balance starting only in the next epoch