[vetoken] past_total_supply (#169)
In order to compute weights, we need a way to be able to query the past
supply at any point. Rather than keeping track of unlock amounts, we can
simplify our algorithm a lot more if we simply keep track of the total
supply per epoch.

When a VeToken is locked, we know which weeks that will result in an
increased supply -- [now_week, end_week - 1]. Rather than an
O(max_locked_weeks) algorithm we had previously, we can have a much
better O(end_week - now_week) which has an upper bound of
`max_locked_weeks`.