[vetoken] delegation support (#206)
`fun delegate_to`. Will create a delegation with the amounts locked for
the applicable epoch. In order to get the total amount delegated. `fun
delegated_balance()` can be used. Like balance/supply past delegated
amounts can be queried for

Important to note that `fun balance()` is unchanged. Even if delegating,
the correct balance is still returned. For consumers which want to
respect delegations -> `fun delegated_balance()`. If self-delegating,
the account's own balance is included

## Aside
- avoid named accounts for non protocol addresses