fix: correct next_claimable
previously, let's say the current epoch is 3, and epoch_dividend is [epoch: 1, dividend: 100] -> [epoch: 2, dividend: 200] -> [epoch: 3, dividend: 50],
after calling `claim`, next_claimable is updated to be 3, while [epoch: 3, dividend: 50] hasn't been claimed and will be skipped on next claim, which is a bug.

after the fix, next_claimable will be updated to be 2 in this case.